+++ avaframe/out3Plot/makePalette.py
